requirements.txt
pytube
pydub
ffmpeg-python
openai-whisper
tqdm